+++ backend/canisters/local_user_index/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Mark channels as read after joining community or channel ([#4004](https://github.com/open-chat-labs/open-chat/pull/4004))
+
 ## [[2.0.752](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.752-local_user_index)] - 2023-07-20
 
 ### Changed